requirements.txt
fastapi[standard]
pydantic
ollama
python-dotenv
websocket-client
pillow